% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_CD_info.R
\name{add_CD_info}
\alias{add_CD_info}
\title{Add experiment information to CircadianData object}
\usage{
add_CD_info(cd, period)
}
\arguments{
\item{cd}{A \code{CircadianData} object.}

\item{period}{A number defining the period length of the circadian data.}
}
\value{
The \code{CircadianData} object with added experiment info
}
\description{
Add experiment information to CircadianData object
}
