% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_arser.R
\name{analyze_arser}
\alias{analyze_arser}
\title{Detect rhythmicity with ARSER}
\usage{
analyze_arser(cd, method_args = list())
}
\arguments{
\item{cd}{A \code{CircadianData} object.}

\item{method_args}{Additional parameters passed to \code{MetaCycle::meta2d()}}
}
\value{
A data frame with the results of the ARSER analysis.
}
\description{
This function runs rhythmicity detection with ARSER
}
\examples{
data(cw_data)
data(cw_metadata)
cd <- CircadianData(
  dataset = cw_data,
  metadata = cw_metadata,
  colname_sample = "Sample_ID",
  colname_time = "Time",
  colname_group = "Group",
  colname_subject = "Subject_ID"
)
cd <- clockworks:::add_experiment_info(cd)
cd <- clockworks:::estimate_wave_params(cd)
results <- clockworks:::analyze_arser(cd)
head(results$res_formatted)
}
