% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_circan.R
\name{analyze_circan}
\alias{analyze_circan}
\title{Detect rhythmicity with CircaN}
\usage{
analyze_circan(cd, method_args = list())
}
\arguments{
\item{cd}{A \code{CircadianData} object.}

\item{method_args}{Additional parameters passed to \code{CircaN::circan()}}
}
\value{
A data frame with the results of the CircaN analysis.
}
\description{
This function runs rhythmicity detection using CircaN.
}
