% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_diffcircadian.R
\name{analyze_diffcircadian}
\alias{analyze_diffcircadian}
\title{Detect rhythmicity with diffCircadian}
\usage{
analyze_diffcircadian(cd, method_args = list())
}
\arguments{
\item{cd}{A \code{CircadianData} object.}

\item{method_args}{Additional parameters passed to \verb{<method_function>}}
}
\value{
A list with the original and formatted results of the diffCircadian analysis.
}
\description{
This function runs rhythmicity detection with diffCircadian
}
