% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_jtkcycle.R
\name{analyze_jtkcycle}
\alias{analyze_jtkcycle}
\title{Detect rhythmicity with JTK_CYCLE}
\usage{
analyze_jtkcycle(cd, method_args = list())
}
\arguments{
\item{cd}{A \code{CircadianData} object.}

\item{method_args}{Additional parameters passed to \verb{<method_function>}}
}
\value{
A data frame with the results of the JTK_CYCLE analysis.
}
\description{
This function runs rhythmicity detection JTK_CYCLE
}
\examples{
data(cw_data)
data(cw_metadata)
cd <- CircadianData(
  dataset = cw_data,
  metadata = cw_metadata,
  colname_sample = "Sample_ID",
  colname_time = "Time",
  colname_group = "Group",
  colname_subject = "Subject_ID"
)
cd <- clockworks:::add_experiment_info(cd)
cd <- clockworks:::estimate_wave_params(cd)
results <- clockworks:::analyze_jtkcycle(cd)
head(results)
}
