% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_limorhyde_gammel.R
\name{analyze_limorhyde_gammel}
\alias{analyze_limorhyde_gammel}
\title{Detect rhythmicity with LimoRhyde}
\usage{
analyze_limorhyde_gammel(cd, method_args)
}
\arguments{
\item{cd}{A \code{CircadianData} object.}

\item{method_args}{Additional parameters passed to \code{limorhyde::limorhyde()}}
}
\value{
A data frame with the results of the LimoRhyde analysis.
}
\description{
This function runs rhythmicity detection with LimoRhyde
}
