% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_ls.R
\name{analyze_ls}
\alias{analyze_ls}
\title{Detect rhythmicity with Lomb-Scargle}
\usage{
analyze_ls(cd, method_args = list())
}
\arguments{
\item{cd}{A \code{CircadianData} object.}

\item{method_args}{Additional parameters passed to \verb{<method_function>}}
}
\value{
A data frame with the results of the Lomb-Scargle analysis.
}
\description{
This function runs rhythmicity detection Lomb-Scargle
}
