% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_meta3d.R
\name{analyze_meta3d}
\alias{analyze_meta3d}
\title{Detect rhythmicity with meta3d}
\usage{
analyze_meta3d(cd, method_args = list())
}
\arguments{
\item{cd}{A \code{CircadianData} object.}

\item{method_args}{Additional parameters passed to \code{MetaCycle::meta3d()}}
}
\value{
A data frame with the results of the meta3d analysis.
}
\description{
This function runs rhythmicity detection with meta3d
}
