% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_metacycle.R
\name{analyze_metacycle}
\alias{analyze_metacycle}
\title{Detect rhythmicity with MetaCycle}
\usage{
analyze_metacycle(cd, method_args = list())
}
\arguments{
\item{cd}{A \code{CircadianData} object.}

\item{method_args}{Additional parameters passed to \code{MetaCycle::meta3d()}}
}
\value{
A list with the results of the MetaCycle analysis
}
\description{
This function calls either \code{MetaCycle::meta2d()} or \code{MetaCycle::meta3d()} to
run rhythmicity detection, depending on whether the data includes repeated
measures or not.
}
