% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_repeatedcircadian.R
\name{analyze_repeatedcircadian}
\alias{analyze_repeatedcircadian}
\title{Detect rhythmicity with RepeatedCircadian}
\usage{
analyze_repeatedcircadian(cd, method_args = list())
}
\arguments{
\item{cd}{A \code{CircadianData} object.}

\item{method_args}{Additional parameters passed to
\code{RepeatedCircadian::rpt_rhythmicity()}}
}
\value{
A list containing the original as and cleaned results of the
RepeatedCircadian analysis.
}
\description{
This function runs rhythmicity detection using RepeatedCircadian
}
\examples{
data(cw_data)
data(cw_metadata)
cw_metadata <- clockworks::check_metadata(
  cw_metadata,
  colname_sample = "Sample_ID",
  colname_time = "Time",
  colname_group = "Group",
  colname_subject = "Subject_ID"
)
cd <- CircadianData(cw_data, cw_metadata)
cd <- clockworks:::add_experiment_info(cd, period = 24)
results <- clockworks:::analyze_repeatedcircadian(cd)
head(results$res_original)
}
