% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_timecycle.R
\name{analyze_timecycle}
\alias{analyze_timecycle}
\title{Detect rhythmicity with TimeCycle}
\usage{
analyze_timecycle(cd, method_args = list())
}
\arguments{
\item{cd}{A \code{CircadianData} object.}

\item{method_args}{Additional parameters passed to \verb{<TimeCycle::TimeCycle()>}}
}
\value{
A list with the original and formatted results of the TimeCycle
analysis.
}
\description{
This function runs rhythmicity detection with TimeCycle
}
