% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CircadianData.R
\name{cd_results}
\alias{cd_results}
\alias{cd_results<-}
\title{#' @rdname CircadianData-accessors
#' @export
setGeneric("results", function(x) standardGeneric("results"))}
\usage{
cd_results(x)

cd_results(x) <- value
}
\arguments{
\item{x}{A \code{CircadianData} object.}

\item{value}{A list to assign as results.}
}
\value{
A list (\code{cd_results}) or a modified \code{CircadianData} object (\verb{cd_results<-}).
}
\description{
#' @rdname CircadianData-accessors
setMethod("results", "CircadianData", function(x) x@results)
}
\details{
#' @rdname CircadianData-accessors
#' @export
setGeneric("results<-", function(x, value) standardGeneric("results<-"))

#' @rdname CircadianData-accessors
setReplaceMethod("results", "CircadianData", function(x, value) {
if (!is.list(value)) stop("'value' must be a list.")
x@results <- value
validObject(x)
x
})
Access or modify the results of a CircadianData object
}
