% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_arser.R
\name{check_arser}
\alias{check_arser}
\title{ARSER CircadianData check}
\usage{
check_arser(cd)
}
\arguments{
\item{cd}{A \code{CircadianData} object}
}
\value{
A \code{CircadianData} object
}
\description{
This function checks whether a \code{CircadianData} object contains all the meta
data columns necessary to run rhythmicity detection with ARSER.
}
