% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_circan.R
\name{check_circan}
\alias{check_circan}
\title{CircaN CircadianData check}
\usage{
check_circan(cd)
}
\arguments{
\item{cd}{A \code{CircadianData} object}
}
\value{
A \code{CircadianData} object
}
\description{
This function checks whether a \code{CircadianData} object contains all the meta
data columns necessary to run rhythmicity detection with \code{CircaN}.
}
