% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_dryr.R
\name{check_dryr}
\alias{check_dryr}
\title{dryR CircadianData check}
\usage{
check_dryr(cd)
}
\arguments{
\item{cd}{A \code{CircadianData} object}
}
\value{
A \code{CircadianData} object
}
\description{
This function checks whether a \code{CircadianData} object contains all the meta
data columns necessary to run rhythmicity detection with dryR.
}
