% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_genecycle.R
\name{check_genecycle}
\alias{check_genecycle}
\title{GeneCycle CircadianData check}
\usage{
check_genecycle(cd)
}
\arguments{
\item{cd}{A \code{CircadianData} object}
}
\value{
A \code{CircadianData} object
}
\description{
This function checks whether a \code{CircadianData} object contains all the meta
data columns necessary to run rhythmicity detection with GeneCycle.
}
