% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_jtkcycle.R
\name{check_jtkcycle}
\alias{check_jtkcycle}
\title{JTK_CYCLE CircadianData check}
\usage{
check_jtkcycle(cd)
}
\arguments{
\item{cd}{A \code{CircadianData} object}
}
\value{
A \code{CircadianData} object
}
\description{
This function checks whether a \code{CircadianData} object contains all the meta
data columns necessary to run rhythmicity detection with JTK_CYCLE.
}
