% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_limorhyde.R
\name{check_limorhyde}
\alias{check_limorhyde}
\title{LimoRhyde CircadianData check}
\usage{
check_limorhyde(cd)
}
\arguments{
\item{cd}{A \code{CircadianData} object}
}
\value{
A \code{CircadianData} object
}
\description{
This function checks whether a \code{CircadianData} object contains all the meta
data columns necessary to run rhythmicity detection with LimoRhyde.
}
