% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_ls.R
\name{check_ls}
\alias{check_ls}
\title{Lomb-Scargle CircadianData check}
\usage{
check_ls(cd)
}
\arguments{
\item{cd}{A \code{CircadianData} object}
}
\value{
A \code{CircadianData} object
}
\description{
This function checks whether a \code{CircadianData} object contains all the meta
data columns necessary to run rhythmicity detection with Lomb-Scargle.
}
