% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_meta2d.R
\name{check_meta2d}
\alias{check_meta2d}
\title{meta2d CircadianData check}
\usage{
check_meta2d(cd)
}
\arguments{
\item{cd}{A \code{CircadianData} object}
}
\value{
A \code{CircadianData} object
}
\description{
This function checks whether a \code{CircadianData} object contains all the meta
data columns necessary to run rhythmicity detection with meta2d.
}
