% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_meta3d.R
\name{check_meta3d}
\alias{check_meta3d}
\title{meta3d CircadianData check}
\usage{
check_meta3d(cd)
}
\arguments{
\item{cd}{A \code{CircadianData} object}
}
\value{
A \code{CircadianData} object
}
\description{
This function checks whether a \code{CircadianData} object contains all the meta
data columns necessary to run rhythmicity detection with meta3d.
}
