% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_metadata.R
\name{check_metadata}
\alias{check_metadata}
\title{Check Metadata for Analysis}
\usage{
check_metadata(
  metadata,
  colname_sample,
  colname_time,
  colname_group = NULL,
  colname_subject = NULL
)
}
\arguments{
\item{metadata}{A data frame containing metadata associated with the samples
in the dataset.}

\item{colname_sample}{A string specifying the name of the column in
\code{metadata} that contains the sample IDs. These IDs should be unique.}

\item{colname_time}{A string specifying the name of the column in \code{metadata}
that contains the time information (e.g., collection time). This column
must be numeric.}

\item{colname_group}{A string (optional) specifying the name of the column in
\code{metadata} that contains group information (e.g., treatment groups,
conditions). If no group information is available, leave as \code{NULL}
(default).}

\item{colname_subject}{A string (optional) specifying the name of the column
in \code{metadata} that contains subject IDs. This is useful for repeated
measures designs where each subject has multiple samples taken over time.
If no subject information is available, leave as \code{NULL} (default).}
}
\value{
A data frame with the metadata, after performing the checks and
column ordering. The columns will be ordered with \code{colname_sample} first,
followed by \code{colname_time}, \code{colname_subject}, \code{colname_group}, and any
additional columns.
}
\description{
This function performs a series of checks on a metadata data frame to ensure
it is suitable for downstream analysis. It verifies the presence and type of
required columns, checks for uniqueness of sample IDs issues warnings
about ignored columns, and renames columns to ensure consistency.
}
\details{
The function performs the following checks and operations:
\enumerate{
\item Checks if \code{metadata} is a data frame.
\item Checks if the required columns specified by \code{colname_time} and
\code{colname_sample} are present in the \code{metadata} data frame.
\item Checks if the optional columns specified by \code{colname_group} and
\code{colname_subject} are present (if they are not \code{NULL}) and contain
reasonable values.
\item Checks if the \code{colname_time} column contains numeric values.
\item Checks if the values in the \code{colname_sample} column are unique.
\item Issues a warning if there are additional columns in \code{metadata} that
are not explicitly used by the function, informing the user that these
columns will be ignored.
\item Converts the metadata to a standard \code{data.frame} (e.g. not a tibble)
\item Adds the sample IDs as row names.
\item Adds columns with pre-defined names and order to make downstream
analysis easier.
}
}
\examples{
# Create dummy metadata
metadata <- data.frame(
  sample_id = paste0("sample", 1:10),
  time = 1:10,
  group = rep(c("A", "B"), each = 5),
  subject = rep(1:2, 5),
  extra_col = rnorm(10)
)

# Check the metadata
checked_metadata <- check_metadata(
  metadata = metadata,
  colname_time = "time",
  colname_sample = "sample_id",
  colname_group = "group",
  colname_subject = "subject"
)

# Example with no group or subject information
checked_metadata_no_group <- check_metadata(
  metadata = metadata,
  colname_time = "time",
  colname_sample = "sample_id"
)

# Example with missing columns
result <- try(check_metadata(
  metadata = metadata[, c("time", "group", "subject")],
  colname_time = "time",
  colname_sample = "sample_id" # This will cause an error
))
if(inherits(result, "try-error")) {
  message("Error caught (as expected).")
}

# Example with non-numeric time
metadata_non_numeric_time <- metadata
metadata_non_numeric_time$time <- letters[1:10]
result <- try(check_metadata(
  metadata = metadata_non_numeric_time,
  colname_time = "time",
  colname_sample = "sample_id"
))
if(inherits(result, "try-error")) {
  message("Error caught (as expected).")
}

# Example with duplicated sample IDs
metadata_duplicated_ids <- metadata
metadata_duplicated_ids$sample_id[1] <- metadata_duplicated_ids$sample_id[2]
result <- try(check_metadata(
  metadata = metadata_duplicated_ids,
  colname_time = "time",
  colname_sample = "sample_id"
))
if(inherits(result, "try-error")) {
  message("Error caught (as expected).")
}

}
