% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_rain.R
\name{check_rain}
\alias{check_rain}
\title{RAIN CircadianData check}
\usage{
check_rain(cd)
}
\arguments{
\item{cd}{A \code{CircadianData} object}
}
\value{
A \code{CircadianData} object
}
\description{
This function checks whether a \code{CircadianData} object contains all the meta
data columns necessary to run rhythmicity detection with RAIN.
}
