% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_repeatedcircadian.R
\name{check_repeatedcircadian}
\alias{check_repeatedcircadian}
\title{RepeatedCircadian CircadianData check}
\usage{
check_repeatedcircadian(cd)
}
\arguments{
\item{cd}{A \code{CircadianData} object}
}
\value{
A \code{CircadianData} object
}
\description{
This function checks whether a \code{CircadianData} object contains all the meta
data columns necessary to run rhythmicity detection with RepeatedCircadian.
}
