% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_timecycle.R
\name{check_timecycle}
\alias{check_timecycle}
\title{TimeCycle CircadianData check}
\usage{
check_timecycle(cd)
}
\arguments{
\item{cd}{A \code{CircadianData} object}
}
\value{
A \code{CircadianData} object
}
\description{
This function checks whether a \code{CircadianData} object contains all the meta
data columns necessary to run rhythmicity detection with TimeCycle.
}
