% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CircadianData.R
\name{dataset}
\alias{dataset}
\alias{metadata}
\alias{experimentInfo}
\alias{experimentInfo<-}
\title{Get the Dataset Matrix}
\usage{
dataset(x)

metadata(x)

experimentInfo(x)

experimentInfo(x) <- value
}
\arguments{
\item{x}{A \code{CircadianData} object.}

\item{value}{A list to replace the current experimentInfo.}
}
\value{
The dataset matrix (features x samples).

The metadata data.frame (samples x attributes).

The experimentInfo list.

The modified \code{CircadianData} object.
}
\description{
Get the Dataset Matrix

Get the Metadata Data Frame

Get the Experiment Information List

Set the Experiment Information List
}
