% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CircadianData.R
\name{CircadianData-subset-expinfo}
\alias{CircadianData-subset-expinfo}
\alias{[[,CircadianData,character,missing-method}
\alias{[[<-,CircadianData,character,missing-method}
\alias{$,CircadianData-method}
\alias{$<-,CircadianData-method}
\title{Access and modify elements of \code{experiment_info}}
\usage{
\S4method{[[}{CircadianData,character,missing}(x, i, j, ...)

\S4method{[[}{CircadianData,character,missing}(x, i, j) <- value

\S4method{$}{CircadianData}(x, name)

\S4method{$}{CircadianData}(x, name) <- value
}
\arguments{
\item{x}{A \code{CircadianData} object.}

\item{i, name}{A single character string specifying the name of the element.}

\item{j}{Not used in these methods.}

\item{...}{Not used}

\item{value}{The value to assign to the element.}
}
\description{
Provides methods for getting and setting elements within the
\code{experiment_info} list slot of a \code{CircadianData} object using standard R
operators.
}
\details{
These methods allow for intuitive interaction with the
\code{experiment_info} list:
\itemize{
\item \code{x$name} or \code{x[["name"]]}: Retrieve an element.
\item \code{x$name <- value} or \code{x[["name"]] <- value}: Assign or add an element.
}
}
\examples{
data("cw_data", package = "clockworks")
data("cw_metadata", package = "clockworks")
cd_obj <- CircadianData(cw_data,
                        cw_metadata,
                        colname_sample = "Sample_ID",
                        colname_time = "Time")

# Add a new element to experiment_info using `$`
cd_obj$period <- 24

# Add another element using `[[`
cd_obj[["data_type"]] <- "norm"

# Retrieve the information
cd_obj$period
cd_obj[["data_type"]]

# See the entire updated list
experiment_info(cd_obj)
}
