% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clockworks.R
\name{clockworks}
\alias{clockworks}
\title{clockworks}
\usage{
clockworks(
  dataset,
  metadata,
  colname_time,
  colname_sample,
  colname_group = NULL,
  colname_subject = NULL,
  period = 24,
  method,
  data_type = "norm",
  method_args = list()
)
}
\arguments{
\item{dataset}{A data frame or matrix representing the dataset to be checked.
It is expected to contain numeric data.}

\item{metadata}{A data frame containing metadata associated with the samples
in the dataset.}

\item{colname_time}{A string specifying the name of the column in \code{metadata}
that contains the time information (e.g., collection time). This column
must be numeric.}

\item{colname_sample}{A string specifying the name of the column in
\code{metadata} that contains the sample IDs. These IDs should be unique.}

\item{colname_group}{A string (optional) specifying the name of the column in
\code{metadata} that contains group information (e.g., treatment groups,
conditions). If no group information is available, leave as \code{NULL}
(default).}

\item{colname_subject}{A string (optional) specifying the name of the column
in \code{metadata} that contains subject IDs. This is useful for repeated
measures designs where each subject has multiple samples taken over time.
If no subject information is available, leave as \code{NULL} (default).}

\item{period}{A number specifying the period of the samples. Can also be a
vector with two numbers giving the minimum and maximum values for the
period. When using a method that requires a single value for the period but
two numbers are provided, the mean of the two will be used.}

\item{method}{A string specifying the method used to analyse the data. Needs
to be one of c("ARSER", "CircaN", "diffCircadian", "GeneCycle",
"JTK_CYCLE", "RepeatedCircadian", "LimoRhyde", "LS", "meta2d", "RAIN",
"TimeCycle")}

\item{data_type}{Type of data in \code{dataset}. Must be one of "count" (for data
following a negative bionmial distribution) or "norm" for data roughly
following a normal distribution (e.g. log-CPM values).}

\item{method_args}{Additional parameters passed to the selected method.}
}
\value{
A list with two data frames containing the original results
(\code{res_original}) as well as formatted results that follow the same
structure for all methods (\code{res_formatted}).
}
\description{
Rhythmicity analysis using your method of choice.
}
