% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CircadianData.R
\name{colnames,CircadianData-method}
\alias{colnames,CircadianData-method}
\title{Get Column Names (Samples)}
\usage{
\S4method{colnames}{CircadianData}(x)
}
\arguments{
\item{x}{A \code{CircadianData} object.}
}
\value{
Character vector of column names (sample IDs).
}
\description{
Get Column Names (Samples)
}
