% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CircadianData.R
\name{colnames<-,CircadianData-method}
\alias{colnames<-,CircadianData-method}
\title{Set Column Names (Samples)}
\usage{
\S4method{colnames}{CircadianData}(x) <- value
}
\arguments{
\item{x}{A \code{CircadianData} object.}

\item{value}{A character vector of column names (sample IDs).}
}
\value{
The modified \code{CircadianData} object.
}
\description{
Set Column Names (Samples)
}
