% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_to_cpm.R
\name{convert_to_cpm}
\alias{convert_to_cpm}
\title{Convert count matrix in a CircadianData object to CPM / log-CPM}
\usage{
convert_to_cpm(cd, log = TRUE)
}
\arguments{
\item{cd}{A \code{CircadianData} object containing count data in the \code{results} slot and a
\code{norm_factors} vector of length equal to the number of samples.}

\item{log}{Logical, whether to return log2-transformed CPM values. Default is TRUE.}
}
\value{
A \code{CircadianData} object with the dataset slot replaced by the CPM or
log-CPM values.
}
\description{
Calculate counts-per-million (CPM) for the count matrix stored inside a
CircadianData object. Library sizes are adjusted using the sample
normalization factors calculated by edgeR's \code{normLibSizes} which are
stored in the CircadianData object. CPM values are then calculated using
edgeR's \code{cpm} function with the adjusted library sizes.
}
