% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_valid.R
\name{count_valid}
\alias{count_valid}
\title{JTK_CYCLE prep}
\usage{
count_valid(cd)
}
\arguments{
\item{cd}{A \code{CircadianData} object}
}
\value{
A data frame with the number of non-NA values per feature and the number
of time points with at least one non-NA value.
}
\description{
This function is used to check the number of non-NA values per feature in a \code{CircadianData} object.
}
