% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CircadianData.R
\name{dim,CircadianData-method}
\alias{dim,CircadianData-method}
\title{Get Dimensions}
\usage{
\S4method{dim}{CircadianData}(x)
}
\arguments{
\item{x}{A \code{CircadianData} object.}
}
\value{
Integer vector of length 2 (features, samples).
}
\description{
Get Dimensions
}
