% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CircadianData.R
\name{dimnames,CircadianData-method}
\alias{dimnames,CircadianData-method}
\title{Get Dimension Names}
\usage{
\S4method{dimnames}{CircadianData}(x)
}
\arguments{
\item{x}{A \code{CircadianData} object.}
}
\value{
A list containing row names (features) and column names (samples).
}
\description{
Get Dimension Names
}
