% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CircadianData.R
\name{dimnames<-,CircadianData,ANY-method}
\alias{dimnames<-,CircadianData,ANY-method}
\title{Set Dimension Names}
\usage{
\S4method{dimnames}{CircadianData,ANY}(x) <- value
}
\arguments{
\item{x}{A \code{CircadianData} object.}

\item{value}{A list of two character vectors (feature names, sample names).}
}
\value{
The modified \code{CircadianData} object.
}
\description{
Set Dimension Names
}
