% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CircadianData.R
\name{estimate_wave_params}
\alias{estimate_wave_params}
\title{Estimating cosine Wave Parameters}
\usage{
estimate_wave_params(cd_obj, remove_batch_effects = FALSE, logCPM = "auto")
}
\arguments{
\item{cd_obj}{A \code{CircadianData} object}

\item{remove_batch_effects}{Logical. If \code{TRUE}, subject batch effects will be
removed before estimating the wave parameters.}

\item{logCPM}{Logical or "auto". If \code{TRUE}, the data will be converted to
logCPM values before estimating the wave parameters. If "auto", logCPM will
be used if the data type is "count".}
}
\value{
A data frame with estimated cosine wave parameters for every feature.
}
\description{
This function is used to get an estimate for the cosine wave parameters for
each feature using harmonic regression.
}
\details{
The returned parameters correspond to the model \deqn{y = M + A
cos(\frac{2\pi}{T} (t - \varphi))} With \eqn{M} the mesor, \eqn{A} the
amplitude, \eqn{T} the period, \eqn{t} the time and \eqn{\varphi} the phase
in the same units as \eqn{t} (e.g. hours).
}
