% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execute_arser.R
\name{execute_arser}
\alias{execute_arser}
\title{Run ARSER rhythmicity detection}
\usage{
execute_arser(inputs, grp, method_args = list(), missing_data_fraction = 0)
}
\arguments{
\item{inputs}{Inputs for rhythmicity detection function generated by
\code{prepare_arser()}}

\item{grp}{Group that is being analysed}

\item{method_args}{Additional parameters passed to \code{MetaCycle::meta2d()}}

\item{missing_data_fraction}{Fraction of missing data in the dataset.}
}
\value{
A list
}
\description{
Run ARSER rhythmicity detection
}
