% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execute_diffcircadian.R
\name{execute_diffcircadian}
\alias{execute_diffcircadian}
\title{Run diffCircadian rhythmicity detection}
\usage{
execute_diffcircadian(inputs, method_args = list())
}
\arguments{
\item{inputs}{Inputs for rhythmicity detection function generated by
\code{prepare_diffcircadian()}}

\item{method_args}{Additional parameters passed to
\code{diffCircadian::LR_rhythmicity()}}
}
\value{
A data frame
}
\description{
Run diffCircadian rhythmicity detection
}
