% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execute_dryr.R
\name{execute_dryr}
\alias{execute_dryr}
\title{Run dryR rhythmicity detection}
\usage{
execute_dryr(inputs, grp, method_args = list())
}
\arguments{
\item{inputs}{Inputs for rhythmicity detection function generated by
\code{prepare_dryr()}}

\item{method_args, }{Additional parameters passed to \code{dryR::f_24()} or
\code{dryR::dryseq_single}}
}
\value{
A data frame
}
\description{
Run dryR rhythmicity detection
}
