% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execute_meta2d.R
\name{execute_meta2d}
\alias{execute_meta2d}
\title{Run meta2d rhythmicity detection}
\usage{
execute_meta2d(inputs, grp, method_args = list())
}
\arguments{
\item{inputs}{Inputs for rhythmicity detection function generated by
\code{prepare_meta2d()}}

\item{grp}{Group that is being analysed}

\item{method_args}{Additional parameters passed to \code{MetaCycle::meta2d()}}
}
\value{
A list
}
\description{
Run meta2d rhythmicity detection
}
