% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execute_meta3d.R
\name{execute_meta3d}
\alias{execute_meta3d}
\title{Run meta3d rhythmicity detection}
\usage{
execute_meta3d(inputs, method_args = list())
}
\arguments{
\item{inputs}{Inputs for rhythmicity detection function generated by
\code{prepare_meta3d()}}

\item{method_args}{Additional parameters passed to \code{MetaCycle::meta3d()}}
}
\value{
A list
}
\description{
Run meta3d rhythmicity detection
}
