% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execute_rain.R
\name{execute_rain}
\alias{execute_rain}
\title{Run RAIN rhythmicity detection}
\usage{
execute_rain(inputs, grp, method_args = list())
}
\arguments{
\item{inputs}{Inputs for rhythmicity detection function generated by
\code{prepare_rain()}}

\item{...}{Additional parameters passed to \code{rain::rain}}
}
\value{
A data frame
}
\description{
Run RAIN rhythmicity detection
}
