% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execute_rain_parallel.R
\name{execute_rain_parallel}
\alias{execute_rain_parallel}
\title{Run RAIN rhythmicity detection in parallel}
\usage{
execute_rain_parallel(inputs, grp, method_args = list())
}
\arguments{
\item{inputs}{Inputs for rhythmicity detection function generated by
\code{prepare_rain()}}

\item{grp}{Group that is analysed}

\item{method_args}{Additional parameters passed to \code{rain::rain}}
}
\value{
A data frame
}
\description{
Run RAIN rhythmicity detection in parallel
}
