% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execute_repeatedcircadian.R
\name{execute_repeatedcircadian}
\alias{execute_repeatedcircadian}
\title{Run RepeatedCircadian rhythmicity detection}
\usage{
execute_repeatedcircadian(inputs, grp, method_args = list())
}
\arguments{
\item{inputs}{Inputs for rhythmicity detection function generated by
\code{prepare_repeatedcircadian()}}

\item{grp}{Group that is analysed}

\item{method_args}{Additional parameters passed to
\code{RepeatedCircadian::rpt_rhythmicity()}}
}
\value{
A data frame
}
\description{
Run RepeatedCircadian rhythmicity detection
}
