% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_arser.R
\name{format_arser}
\alias{format_arser}
\title{Format ARSER Results}
\usage{
format_arser(
  ls_res_groups,
  ls_harm_groups,
  added_group,
  log_transformed = FALSE,
  log_base = 2
)
}
\arguments{
\item{ls_res_groups}{A list with results from a ARSER rhythmicity analysis,
split into groups.}

\item{ls_harm_groups}{A list with results from a harmonic regression
analysis, split into groups.}

\item{added_group}{If TRUE the "group" column will be removed}

\item{log_transformed}{Logical, if the data is log-transformed}

\item{log_base}{Logarithmic base of data. Only relevant if \code{log_transformed}
is \code{TRUE}.}
}
\value{
A list of data frames containing the original and formatted results
}
\description{
Format ARSER Results
}
