% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_dryr.R
\name{format_dryr}
\alias{format_dryr}
\title{Format dryR Results}
\usage{
format_dryr(ls_res_groups, w_params, method, added_group)
}
\arguments{
\item{ls_res_groups}{A list with results from a dryR rhythmicity analysis,
split into groups.}

\item{w_params}{A data frame with results from a harmonic regression
analysis.}

\item{method}{Which method ("f_24" or "dryseq_single") was used in the
analysis.}

\item{added_group}{If TRUE the "group" column will be removed}
}
\value{
A list of data frames containing the original and formatted results
}
\description{
Format dryR Results
}
