% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_limorhyde.R
\name{format_limorhyde}
\alias{format_limorhyde}
\title{Format LimoRhyde Results}
\usage{
format_limorhyde(df_original, ls_harm_groups, period)
}
\arguments{
\item{df_original}{A data frame with results from a LimoRhyde rhythmicity
analysis.}

\item{ls_harm_groups}{A list with results from a harmonic regression
analysis, split into groups.}

\item{period}{Period from \code{CircadianData} object}
}
\value{
A list of data frames containing the original and formatted results
}
\description{
Format LimoRhyde Results
}
