% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_ls.R
\name{format_ls}
\alias{format_ls}
\title{Format Lomb-Scargle Results}
\usage{
format_ls(ls_res_groups, remove_group)
}
\arguments{
\item{ls_res_groups}{A list with results from a Lomb-Scargle rhythmicity
analysis, split into groups.}

\item{remove_group}{If TRUE the "group" column will be removed}
}
\value{
A list of data frames containing the original and formatted results
}
\description{
Format Lomb-Scargle Results
}
