% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_meta2d.R
\name{format_meta2d}
\alias{format_meta2d}
\title{Format meta2d Results}
\usage{
format_meta2d(ls_res_groups, added_group)
}
\arguments{
\item{ls_res_groups}{A list with results from a meta2d rhythmicity
analysis, split into groups.}

\item{added_group}{If TRUE the "group" column will be removed}
}
\value{
A list of data frames containing the original and formatted results
}
\description{
Format meta2d Results
}
