% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_meta3d.R
\name{format_meta3d}
\alias{format_meta3d}
\title{Format meta3d Results}
\usage{
format_meta3d(ls_res, w_params, meta3d_method)
}
\arguments{
\item{w_params}{A data frame with results from a harmonic regression
analysis.}

\item{meta3d_method}{The method used in the meta3d analysis.}
}
\value{
A list of data frames containing the original and formatted results
}
\description{
Format meta3d Results
}
