% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_repeatedcircadian.R
\name{format_repeatedcircadian}
\alias{format_repeatedcircadian}
\title{Format RepeatedCircadian Results}
\usage{
format_repeatedcircadian(ls_res_groups, added_group)
}
\arguments{
\item{ls_res_groups}{A list with results from a RepeatedCircadian rhythmicity
analysis, split into groups.}

\item{added_group}{If TRUE the "group" column will be removed}
}
\value{
A list of data frames containing the original and formatted results
}
\description{
Format RepeatedCircadian Results
}
