% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_timecycle.R
\name{format_timecycle}
\alias{format_timecycle}
\title{Format TimeCycle Results}
\usage{
format_timecycle(ls_res_groups, ls_harm_groups, resamplings, added_group)
}
\arguments{
\item{ls_res_groups}{A list with results from a TimeCycle rhythmicity
analysis, split into groups.}

\item{ls_harm_groups}{A list with results from a harmonic regression
analysis, split into groups.}

\item{resamplings}{Number of resamplings used to calculate p-values.}

\item{added_group}{If TRUE the "group" column will be removed}
}
\value{
A list of data frames containing the original and formatted results
}
\description{
Format TimeCycle Results
}
