% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CircadianData.R
\name{get_dataset}
\alias{get_dataset}
\title{Extract Data Set from a CircadianData Object}
\usage{
get_dataset(cd)
}
\arguments{
\item{cd}{A \code{CircadianData} object.}
}
\value{
A matrix with features as row names and sample IDs as column names
}
\description{
\code{get_dataset()} retrieves the data set stored inside a \code{CircadianData}
object.
}
