% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CircadianData.R
\name{get_experiment_info}
\alias{get_experiment_info}
\title{Extract Experiment Info from a CircadianData Object}
\usage{
get_experiment_info(cd)
}
\arguments{
\item{cd}{A \code{CircadianData} object.}
}
\value{
A list with all stored experiment info
}
\description{
\code{get_experiment_info()} retrieves the experiment info stored inside a
\code{CircadianData} object.
}
