% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CircadianData.R
\name{get_metadata}
\alias{get_metadata}
\title{Extract Meta Data from a CircadianData Object}
\usage{
get_metadata(cd)
}
\arguments{
\item{cd}{A \code{CircadianData} object.}
}
\value{
A data frame with sample IDs as row names and formatted column names
}
\description{
\code{get_metadata()} retrieves the meta data stored inside a \code{CircadianData}
object.
}
