% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CircadianData.R
\name{get_results}
\alias{get_results}
\title{Extract Results From a CircadianData Object}
\usage{
get_results(cd, method = "all", type = c("formatted", "original", "both"))
}
\arguments{
\item{cd}{A \code{CircadianData} object.}

\item{method}{Character. Either \code{"all"} (default) to extract results for all
methods, or the name of a specific method.}

\item{type}{Character. Type of results to return. One of:
\describe{
\item{\code{"formatted"}}{Return standardised result data frames (default).}
\item{\code{"original"}}{Return the original method outputs.}
\item{\code{"both"}}{Return both for each method.}
}}
}
\value{
If a specific method is selected, returns either:
\itemize{
\item a data frame (\code{type = "formatted"}),
\item a method-specific object (\code{type = "original"}), or
\item a list containing both (\code{type = "both"}).
}

If \code{method = "all"}, returns a named list containing the requested type of
results for each method.
}
\description{
\code{get_results()} retrieves standardised ("formatted") or original results for
one or all rhythm detection methods contained in a \code{CircadianData} object.
}
\examples{
\dontrun{
# Retrieve formatted results for all methods
get_results(cd)

# Get formatted results for one method
get_results(cd, method = "RAIN")

# Get original results for all methods
get_results(cd, type = "original")
}

}
