% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_file.R
\name{load_file}
\alias{load_file}
\title{Read data file}
\usage{
load_file(filename, delim = NULL, ...)
}
\arguments{
\item{filename}{A string containing the path to the meta data file.}

\item{delim}{A single character specifying the field separator. If \code{filename}
points to a \code{.csv} or \code{.tsv} file this defaults to "," or TAB,
respectively. Otherwise, needs to be specified.}

\item{...}{Arguments to be passed to \code{read_delim()}}
}
\value{
A \code{data.frame}.
}
\description{
Reads a file and creates a data frame from it. Tries to infer the delimiter
from the file extension and uses the appropriate function among \code{read_csv()},
\code{read_tsv()} and \code{read_delim()} from the \code{readr} package.
}
\examples{
filename <- system.file(
  "extdata",
  "metadata_example.txt.gz",
  package = "clockworks",
  mustWork = TRUE
)
metadata <- load_file(filename = filename, delim = "\t")
head(metadata)
}
