% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CircadianData.R
\name{ncol,CircadianData-method}
\alias{ncol,CircadianData-method}
\title{Get Number of Columns (Samples)}
\usage{
\S4method{ncol}{CircadianData}(x)
}
\arguments{
\item{x}{A \code{CircadianData} object.}
}
\value{
Integer, number of columns (samples).
}
\description{
Get Number of Columns (Samples)
}
