% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalise_dataset.R
\name{normalise_dataset}
\alias{normalise_dataset}
\title{Turn RNA-seq counts to log2 CPM values}
\usage{
normalise_dataset(counts, group = NULL, verbose = TRUE)
}
\arguments{
\item{counts}{A matrix with RNA-seq count data}

\item{group}{A vector or factor giving the experimental group or treatment
condition for each sample}

\item{verbose}{Logical, whether a message should be printed}
}
\value{
A matrix with log2 CPM values
}
\description{
Turn RNA-seq counts to log2 CPM values
}
