% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CircadianData.R
\name{nrow,CircadianData-method}
\alias{nrow,CircadianData-method}
\title{Get Number of Rows (Features)}
\usage{
\S4method{nrow}{CircadianData}(x)
}
\arguments{
\item{x}{A \code{CircadianData} object.}
}
\value{
Integer, number of rows (features).
}
\description{
Get Number of Rows (Features)
}
