% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CircadianData.R
\name{plot_phase_estimates}
\alias{plot_phase_estimates}
\title{Plot Phase Estimates in a Circular Histogram}
\usage{
plot_phase_estimates(
  cd,
  n_bins = 24,
  n_labels = 4,
  pval_adj_cutoff = 1,
  method = NA,
  initial_offset = 5,
  step_size = 1.25,
  title = NULL,
  add_border = TRUE,
  alpha_range = c(0.05, 1),
  default_colour = "deepskyblue"
)
}
\arguments{
\item{cd}{A \code{CircadianData} object that contains wave parameters
generated by \code{estimate_wave_params()} (ran automatically when the
\code{clockworks} function is used).}

\item{n_bins}{An integer specifying the number of bins to divide the period
into for the histogram. Defaults to 24.}

\item{n_labels}{An integer specifying the number of evenly spaced time labels
to display on the circular axis. Defaults to 4 (e.g., 0/24, 6, 12, 18).}

\item{pval_adj_cutoff}{A numeric value between 0 and 1. Only features with an
adjusted p-value below this cutoff will be included in the plot. Requires a
corresponding analysis result to be stored in the \code{results} slot. Defaults
to 1 (including all features).}

\item{method}{An optional character string specifying which analysis result
from the \code{results} slot to use for p-value filtering. If \code{NA} (default) and
a \code{pval_adj_cutoff} is used, the function will use the first available
analysis result.}

\item{initial_offset}{A numeric value controlling the radius of the innermost
ring from the center of the plot.}

\item{step_size}{A numeric value controlling the distance between concentric
rings when multiple groups are plotted. The height of the individual rings
is fixed at 1.}

\item{title}{An optional character string for the plot's main title.}

\item{default_colour}{A character string specifying the color to use for the
plot when no groups are present.}
}
\value{
A \code{ggplot} object representing the circular phase plot. This can be
printed to display the plot or modified with additional \code{ggplot2} layers.
}
\description{
Creates a circular plot showing the distribution of phase estimates for
features. The plot is structured as concentric rings, with each ring
representing a different group (if groups are present). The density of
features peaking at each time point is visualized using color intensity.
}
\details{
This function visualizes the output of wave-fitting analyses (harmonic
regression) stored in a \code{CircadianData} object. It bins the phase estimates
(peak times) for each feature into discrete time intervals and plots their
density on a circular axis representing the full period (e.g., 24 hours).
Features can be pre-filtered based on statistical significance from an
analysis stored in the \code{results} slot of the object.

The plot is built using \code{ggplot2}, and the returned object can be further
customized with standard \code{ggplot2} functions.
}
