% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_repeatedcircadian.R
\name{prep_repeatedcircadian}
\alias{prep_repeatedcircadian}
\title{Prepare data for analysis with RepeatedCircadian}
\usage{
prep_repeatedcircadian(cd)
}
\arguments{
\item{cd}{A \code{CircadianData} object.}
}
\value{
A data frame.
}
\description{
Merge dataset and metadata of a \code{CircadianData} object.
}
