% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_repeatedcircadian_V1.R
\name{prep_repeatedcircadian_V1}
\alias{prep_repeatedcircadian_V1}
\title{Prepare data for analysis with RepeatedCircadian}
\usage{
prep_repeatedcircadian_V1(cd)
}
\arguments{
\item{cd}{A \code{CircadianData} object.}
}
\value{
A long format data frame
}
\description{
Merge dataset and metadata of a \code{CircadianData} object.
}
