% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_diffcircadian.R
\name{prepare_diffcircadian}
\alias{prepare_diffcircadian}
\title{Prepare data for analysis with diffCircadian}
\usage{
prepare_diffcircadian(cd)
}
\arguments{
\item{cd}{A \code{CircadianData} object.}
}
\value{
A data frame.
}
\description{
Merge dataset and metadata of a \code{CircadianData} object.
}
