% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_dryr.R
\name{prepare_dryr}
\alias{prepare_dryr}
\title{dryR prep}
\usage{
prepare_dryr(cd, grp)
}
\arguments{
\item{cd}{A \code{CircadianData} object}

\item{grp}{A string specifying a value in the "group" column of the metadata
slot of \code{cd} which is used for filtering.}
}
\value{
A list with inputs for \code{execute_dryr()}
}
\description{
This function is used to create the input required to run rhythmicity
detection with dryR.
}
