% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_limorhyde.R
\name{prepare_limorhyde}
\alias{prepare_limorhyde}
\title{LimoRhyde prep}
\usage{
prepare_limorhyde(cd)
}
\arguments{
\item{cd}{A \code{CircadianData} object}
}
\value{
A list with inputs for \code{execute_limorhyde()}
}
\description{
This function is used to create the input required to run rhythmicity
detection with LimoRhyde.
}
