% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_ls.R
\name{prepare_ls}
\alias{prepare_ls}
\title{Lomb-Scargle prep}
\usage{
prepare_ls(cd, grp)
}
\arguments{
\item{cd}{A \code{CircadianData} object}

\item{grp}{A string specifying a value in the ".group" column of the metadata
slot of \code{cd} which is used for filtering.}
}
\value{
A list with inputs for \code{execute_ls()}
}
\description{
This function is used to create the input required to run rhythmicity
detection with Lomb-Scargle.
}
