% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_meta3d.R
\name{prepare_meta3d}
\alias{prepare_meta3d}
\title{meta3d prep}
\usage{
prepare_meta3d(cd, cycMethodOne = "JTK")
}
\arguments{
\item{cd}{A \code{CircadianData} object}

\item{cycMethodOne}{The method to use for rhythmicity detection in meta3d.}
}
\value{
A list with inputs for \code{execute_meta3d()}
}
\description{
This function is used to create the input required to run rhythmicity
detection with meta3d.
}
