% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_repeatedcircadian_long.R
\name{prepare_repeatedcircadian_long}
\alias{prepare_repeatedcircadian_long}
\title{Prepare data for analysis with RepeatedCircadian}
\usage{
prepare_repeatedcircadian_long(cd)
}
\arguments{
\item{cd}{A \code{CircadianData} object.}
}
\value{
A data frame.
}
\description{
Merge dataset and metadata of a \code{CircadianData} object.
}
