% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_counts.R
\name{preprocess_counts}
\alias{preprocess_counts}
\title{Preprocess count data by filtering lowly expressed features and normalising the data}
\usage{
preprocess_counts(cd, filter_counts = TRUE, verbose = TRUE)
}
\arguments{
\item{cd}{A \code{CircadianData} object with count data in the \code{results} slot.}

\item{filter_counts}{A logical value specifying whether lowly expressed
features should be filtered out from count data.}

\item{verbose}{Logical, whether a message should be printed.}
}
\value{
The updated \code{CircadianData} object with preprocessed data in the
\code{results} slot
}
\description{
Uses edgeR's \code{filterByExpr} and \code{normLibSizes} functions to filter lowly expressed
features and normalise the data.
}
