% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CircadianData.R
\name{rownames,CircadianData-method}
\alias{rownames,CircadianData-method}
\title{Get Row Names (Features)}
\usage{
\S4method{rownames}{CircadianData}(x)
}
\arguments{
\item{x}{A \code{CircadianData} object.}
}
\value{
Character vector of row names (feature IDs).
}
\description{
Get Row Names (Features)
}
