% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CircadianData.R
\name{rownames<-,CircadianData-method}
\alias{rownames<-,CircadianData-method}
\title{Set Row Names (Features)}
\usage{
\S4method{rownames}{CircadianData}(x) <- value
}
\arguments{
\item{x}{A \code{CircadianData} object.}

\item{value}{A character vector of row names (feature IDs).}
}
\value{
The modified \code{CircadianData} object.
}
\description{
Set Row Names (Features)
}
