% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_arser.R
\name{run_arser}
\alias{run_arser}
\title{Detect rhythmicity with ARSER}
\usage{
run_arser(cd, ...)
}
\arguments{
\item{cd}{A \code{CircadianData} object.}

\item{...}{Additional parameters passed to \code{MetaCycle::meta2d()}}
}
\value{
A data frame with the results of the ARSER analysis.
}
\description{
This function runs rhythmicity detection using ARSER
}
