% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_circan.R
\name{run_circan}
\alias{run_circan}
\title{Detect rhythmicity with CircaN}
\usage{
run_circan(cd, ...)
}
\arguments{
\item{cd}{A \code{CircadianData} object.}

\item{...}{Additional parameters passed to \code{CircaN::circan()}}
}
\value{
A data frame with the results of the CircaN analysis.
}
\description{
This function runs rhythmicity detection using CircaN.
}
\examples{
data(cw_data)
data(cw_metadata)
cw_metadata <- clockworks::check_metadata(
  cw_metadata,
  colname_sample = "Sample_ID",
  colname_time = "Time",
  colname_group = "Group",
  colname_subject = "Subject_ID"
)
cd <- CircadianData(cw_data, cw_metadata, experiment_info = list(period = 24, repeated_measures = TRUE))
results <- clockworks:::run_circan(cd)
head(results)
}
