% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_genecycle.R
\name{run_genecycle}
\alias{run_genecycle}
\title{Detect rhythmicity with GeneCycle}
\usage{
run_genecycle(cd, ...)
}
\arguments{
\item{cd}{A \code{CircadianData} object.}

\item{...}{Additional parameters passed to ...}
}
\value{
A data frame with the results of the GeneCycle analysis.
}
\description{
This function runs rhythmicity detection using GeneCycle
}
