% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_limorhyde.R
\name{run_limorhyde}
\alias{run_limorhyde}
\title{Title}
\usage{
run_limorhyde(cd, colname_time)
}
\arguments{
\item{cd}{A \code{CircadianData} object.}

\item{colname_time}{A string specifying the name of the column in the meta
data of \code{cd} that contains the time information.}
}
\value{
A data frame with the results of a LimoRhyde analysis.
}
\description{
Title
}
