% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_meta2d.R
\name{run_meta2d}
\alias{run_meta2d}
\title{Detect rhythmicity with meta2d}
\usage{
run_meta2d(cd, ...)
}
\arguments{
\item{cd}{A \code{CircadianData} object.}

\item{...}{Additional parameters passed to \code{MetaCycle::meta2d()}}
}
\value{
A data frame with the results of the meta2d analysis.
}
\description{
This function runs rhythmicity detection using meta2d
}
