% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_repeatedcircadian.R
\name{run_repeatedcircadian}
\alias{run_repeatedcircadian}
\title{Detect rhythmicity with RepeatedCircadian}
\usage{
run_repeatedcircadian(cd, ...)
}
\arguments{
\item{cd}{A \code{CircadianData} object.}

\item{...}{Additional parameters passed to \code{RepeatedCircadian::rpt_rhythmicity()}}
}
\value{
A data frame with the results of the RepeatedCircadian analysis.
}
\description{
This function runs rhythmicity detection using RepeatedCircadian
}
\examples{
data(cw_data)
data(cw_metadata)
cw_metadata <- clockworks:::check_metadata(
  cw_metadata,
  colname_sample = "Sample_ID",
  colname_time = "Time",
  colname_group = "Group",
  colname_subject = "Subject_ID"
)
cd <- CircadianData(cw_data,
                    cw_metadata,
                    experiment_info = list(
                      period = 24,
                      repeated_measures = TRUE,
                      n_groups = 2
                    ))
#results <- clockworks:::run_repeatedcircadian(cd)
#head(results)
}
