% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CircadianData.R
\name{show,CircadianData-method}
\alias{show,CircadianData-method}
\title{Show Method for CircadianData}
\usage{
\S4method{show}{CircadianData}(object)
}
\arguments{
\item{object}{A \code{CircadianData} object.}
}
\description{
Provides a comprehensive summary of the CircadianData object, including
dimensions, snippets of feature/sample names, previews of the metadata
and dataset, and the contents of the experiment_info list.
}
