% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_exp_info.R
\name{validate_exp_info}
\alias{validate_exp_info}
\title{Validate experiment info before analysis}
\usage{
validate_exp_info(cd_obj)
}
\arguments{
\item{cd_obj}{A \code{CircadianData} object}
}
\description{
Validate experiment info before analysis
}
