% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_experiment_info.R
\name{add_experiment_info}
\alias{add_experiment_info}
\title{Add experiment information to CircadianData object}
\usage{
add_experiment_info(
  cd,
  period = NULL,
  data_type = NULL,
  log_transformed = NULL,
  log_base = NULL,
  estimate_delta_t = TRUE
)
}
\arguments{
\item{cd}{A \code{CircadianData} object.}

\item{period}{A number defining the period length of the circadian data.}

\item{data_type}{Type of data in \code{dataset} slot of \code{CircadianData} object.
Must be one of "count" (for data following a negative bionmial
distribution) or "norm" for data roughly following a normal distribution
(e.g. log-CPM values).}

\item{log_transformed}{Logical, whether or not the data is log-transformed.}

\item{log_base}{A number specifying the logarithmic base of the
log-transformed data. Only relevant if \code{log_transformed} is \code{TRUE}.}

\item{estimate_delta_t}{If TRUE, sampling interval will be estimated from
meta data.}
}
\value{
The \code{CircadianData} object with added experiment info
}
\description{
Add experiment information to CircadianData object
}
