% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_dryr.R
\name{analyze_dryr}
\alias{analyze_dryr}
\title{Detect rhythmicity with dryR}
\usage{
analyze_dryr(cd, method_args = list())
}
\arguments{
\item{cd}{A \code{CircadianData} object.}

\item{method_args}{Additional parameters passed to \code{dryR::f_24()} or
\code{dryR::dryseq_single()}}
}
\value{
A list with the original and formatted results of the dryR analysis.
}
\description{
This function runs rhythmicity detection with dryR
}
\examples{
data(cw_data)
data(cw_metadata)
cw_metadata <- clockworks::check_metadata(
  cw_metadata,
  colname_sample = "Sample_ID",
  colname_time = "Time",
  colname_group = "Group",
  colname_subject = "Subject_ID"
)
cd <- CircadianData(cw_data, cw_metadata)
cd <- clockworks:::add_experiment_info(cd, period = 24, data_type = "norm")
results <- clockworks:::analyze_dryr(cd)
head(results)
}
