% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_genecycle.R
\name{analyze_genecycle}
\alias{analyze_genecycle}
\title{Detect rhythmicity with GeneCycle}
\usage{
analyze_genecycle(cd, method_args = list())
}
\arguments{
\item{cd}{A \code{CircadianData} object.}

\item{method_args}{Additional parameters passed to \code{GeneCycle:robust.g.test()}}
}
\value{
A list with the original and formatted results of the GeneCycle analysis.
}
\description{
This function runs rhythmicity detection with GeneCycle
}
