% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_jtkcycle.R
\name{analyze_jtkcycle}
\alias{analyze_jtkcycle}
\title{Detect rhythmicity with JTK_CYCLE}
\usage{
analyze_jtkcycle(cd, method_args = list())
}
\arguments{
\item{cd}{A \code{CircadianData} object.}

\item{method_args}{Additional parameters passed to \verb{<method_function>}}
}
\value{
A data frame with the results of the JTK_CYCLE analysis.
}
\description{
This function runs rhythmicity detection JTK_CYCLE
}
