% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_jtkcycle.R
\name{analyze_jtkcycle}
\alias{analyze_jtkcycle}
\title{Detect rhythmicity with JTK_CYCLE}
\usage{
analyze_jtkcycle(cd, ...)
}
\arguments{
\item{cd}{A \code{CircadianData} object.}

\item{...}{Additional parameters passed to \code{MetaCycle::meta2d()}}
}
\value{
A data frame with the results of the JTK_CYCLE analysis.
}
\description{
This function runs rhythmicity detection using JTK_CYCLE
}
