% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_limorhyde.R
\name{analyze_limorhyde}
\alias{analyze_limorhyde}
\title{Detect rhythmicity with LimoRhyde}
\usage{
analyze_limorhyde(cd, method_args = list())
}
\arguments{
\item{cd}{A \code{CircadianData} object.}

\item{method_args}{Additional parameters passed to \code{limma::lmfit()} or
\code{edgeR::voomLmFit()}}
}
\value{
A list with the original and formatted results of the LimoRhyde analysis.
}
\description{
This function runs rhythmicity detection with LimoRhyde
}
\examples{
data(cw_data)
data(cw_metadata)
cd <- CircadianData(
  dataset = cw_data,
  metadata = cw_metadata,
  colname_sample = "Sample_ID",
  colname_time = "Time",
  colname_group = "Group",
  colname_subject = "Subject_ID"
)
cd <- clockworks:::add_experiment_info(cd)
cd <- clockworks:::estimate_wave_params(cd)
results <- clockworks:::analyze_limorhyde(cd)
head(results)
}
