% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_limorhyde.R
\name{analyze_limorhyde}
\alias{analyze_limorhyde}
\title{Detect rhythmicity with LimoRhyde}
\usage{
analyze_limorhyde(cd, method_args = list())
}
\arguments{
\item{cd}{A \code{CircadianData} object.}

\item{method_args}{Additional parameters passed to \code{limma::lmfit()} or
\code{edgeR::voomLmFit()}. If included, \code{robust} is passed to \code{limma::eBayes()}.}
}
\value{
A list with the original and formatted results of the LimoRhyde analysis.
}
\description{
This function runs rhythmicity detection with LimoRhyde
}
