% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_limorhyde.R
\name{analyze_limorhyde}
\alias{analyze_limorhyde}
\title{Detect rhythmicity with LimoRhyde}
\usage{
analyze_limorhyde(cd, method_args = list())
}
\arguments{
\item{cd}{A \code{CircadianData} object.}

\item{method_args}{Additional parameters passed to \verb{<method_function>}}
}
\value{
A list with the original and formatted results of the LimoRhyde analysis.
}
\description{
This function runs rhythmicity detection with LimoRhyde
}
\examples{
data(cw_data)
data(cw_metadata)
cw_metadata <- clockworks::check_metadata(
  cw_metadata,
  colname_sample = "Sample_ID",
  colname_time = "Time",
  colname_group = "Group",
  colname_subject = "Subject_ID"
)
cd <- CircadianData(cw_data, cw_metadata)
cd <- clockworks:::add_experiment_info(cd, period = 24, data_type = "norm")
results <- clockworks:::analyze_limorhyde(cd)
head(results)
}
