% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_meta2d.R
\name{analyze_meta2d}
\alias{analyze_meta2d}
\title{Detect rhythmicity with meta2d}
\usage{
analyze_meta2d(cd, method_args = list())
}
\arguments{
\item{cd}{A \code{CircadianData} object.}

\item{method_args}{Additional parameters passed to \code{MetaCycle::meta2d()}}
}
\value{
A data frame with the results of the meta2d analysis.
}
\description{
This function runs rhythmicity detection with meta2d
}
