% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_rain.R
\name{analyze_rain}
\alias{analyze_rain}
\title{Detect rhythmicity with RAIN}
\usage{
analyze_rain(cd, method_args = list())
}
\arguments{
\item{cd}{A \code{CircadianData} object.}

\item{method_args}{Additional parameters passed to \verb{<method_function>}}
}
\value{
A list with the original and formatted results of the RAIN analysis.
}
\description{
This function runs rhythmicity detection with RAIN
}
