% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clockworks.R
\name{clockworks}
\alias{clockworks}
\title{Streamlined Rhythmicity Analysis of Time Series Data}
\usage{
clockworks(cd, method, method_args = list())
}
\arguments{
\item{cd}{A \code{CircadianData} object with added experiment info.}

\item{method}{A string specifying the rhythmicity detection method to use.
Supported methods include "ARSER", "CircaN", "diffCircadian", "dryR",
"GeneCycle", "JTK_CYCLE", "RepeatedCircadian", "LimoRhyde", "LS"
(Lomb-Scargle), "meta2d", "RAIN", and "TimeCycle".}

\item{method_args}{A list of additional arguments to be passed directly to
the chosen analysis method. This allows for more advanced customisation of
the analysis.}
}
\value{
A \code{CircadianData} object including the results of a rhythmicity
analysis using the chosen \code{method}.
}
\description{
The main \code{clockworks()} function is used to detect rhythms in time series
data using a variety of rhythmicity analysis methods. The function takes care
of the data wrangling and formatting internally, allowing to use different
methods by using the same input and changing just one parameter.
}
\details{
Additional details coming soon...
}
