% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cw_data}
\alias{cw_data}
\alias{cw_metadata}
\title{Example rhythmic data}
\format{
\subsection{\code{cw_data}}{

A matrix with 10 rows and 96 columns
}

\subsection{\code{cw_metadata}}{

A data frame with 96 rows and 4 columns
\describe{
\item{Sample_ID}{Sample IDs, correspond to the row names of \code{cw_data}}
\item{Time}{Sample collection time}
\item{Group}{Experimental group / condition}
\item{Subject_ID}{Subject IDs to mark repeated measures}
}
}
}
\usage{
cw_data

cw_metadata
}
\description{
This data set is an example for what the \code{dataset} and \code{metadata} arguments
for the \code{clockworks} function could look like. It contains log-CPM values for
10 genes. Of these, the first two are rhythmic while the remaining eight are
just noise. This synthetic data set contains repeated measures of four
individuals (S1 - S4) that are split into two groups ("A" and "B"), sampled
over a span of 48 hours at 2 hour intervals.
}
\keyword{datasets}
