% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CircadianData.R
\name{estimate_wave_params}
\alias{estimate_wave_params}
\title{Estimating sine Wave Parameters}
\usage{
estimate_wave_params(cd_obj)
}
\arguments{
\item{cd_obj}{A \code{CircadianData} object}
}
\value{
A data frame with estimated sine wave parameters for every feature.
}
\description{
This function is used to get an estimate for the amplitude and phase using
harmonic regression.
}
\details{
The returned parameters correspond to the model \deqn{y = M + A
cos(\frac{2\pi}{T} (t - \phi))} With \eqn{M} the mesor, \eqn{A} the
amplitude, \eqn{T} the period, \eqn{t} the time and \eqn{\phi} the phase in
the same units as \eqn{t} (e.g. hours).
}
