% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execute_genecycle.R
\name{execute_genecycle}
\alias{execute_genecycle}
\title{Run GeneCycle rhythmicity detection}
\usage{
execute_genecycle(inputs, grp, method_args = list())
}
\arguments{
\item{inputs}{Inputs for rhythmicity detection function generated by
\code{prepare_genecycle()}}

\item{grp}{Group that is analysed}

\item{method_args, }{Additional parameters passed to
\code{GeneCycle::robust.spectrum()} and \code{GeneCycle::robust.g.test()}}
}
\value{
A data frame
}
\description{
Run GeneCycle rhythmicity detection
}
