% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execute_jtkcycle.R
\name{execute_jtkcycle}
\alias{execute_jtkcycle}
\title{Run JTK_CYCLE rhythmicity detection}
\usage{
execute_jtkcycle(inputs, grp, method_args = list())
}
\arguments{
\item{inputs}{Inputs for rhythmicity detection function generated by
\code{prepare_jtkcycle()}}

\item{grp}{Group that is analysed}

\item{method_args}{Additional parameters passed to \code{MetaCycle::meta2d()}}
}
\value{
A list
}
\description{
Run JTK_CYCLE rhythmicity detection
}
