% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execute_limorhyde.R
\name{execute_limorhyde}
\alias{execute_limorhyde}
\title{Run LimoRhyde rhythmicity detection}
\usage{
execute_limorhyde(inputs, groups, method_args = list())
}
\arguments{
\item{inputs}{Inputs for rhythmicity detection function generated by
\code{prepare_limorhyde()}.}

\item{groups}{A vector with the groups in the experiment.}

\item{method_args}{Additional parameters passed to \code{limma::lmFit()} or
\code{edgeR::vommLmFit()} depending on \code{type}.}
}
\value{
A data frame
}
\description{
Run LimoRhyde rhythmicity detection
}
