% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execute_ls.R
\name{execute_ls}
\alias{execute_ls}
\title{Run Lomb-Scargle rhythmicity detection}
\usage{
execute_ls(inputs, ...)
}
\arguments{
\item{inputs}{Inputs for rhythmicity detection function generated by
\code{prepare_ls()}}

\item{...}{Additional parameters passed to \code{MetaCycle::meta2d()}}
}
\value{
A list
}
\description{
Run Lomb-Scargle rhythmicity detection
}
