% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execute_timecycle.R
\name{execute_timecycle}
\alias{execute_timecycle}
\title{Run TimeCycle rhythmicity detection}
\usage{
execute_timecycle(inputs, grp, method_args = list())
}
\arguments{
\item{inputs}{Inputs for rhythmicity detection function generated by
\code{prepare_timecycle()}}

\item{grp}{Group that is analysed}

\item{method_args}{Additional parameters passed to \code{TimeCycle::TimeCycle()}}
}
\value{
A data frame
}
\description{
Run TimeCycle rhythmicity detection
}
