% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_diffcircadian.R
\name{format_diffcircadian}
\alias{format_diffcircadian}
\title{Format diffCircadian Results}
\usage{
format_diffcircadian(res_original, period, added_group)
}
\arguments{
\item{res_original}{A data frame with results from a diffCircadian rhythmicity
analysis.}

\item{period}{Period entered by the user.}

\item{remove_group}{If TRUE the "group" column will be removed.}
}
\value{
A list of data frames containing the original and formatted results
}
\description{
Format diffCircadian Results
}
