% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_genecycle.R
\name{format_genecycle}
\alias{format_genecycle}
\title{Format GeneCycle Results}
\usage{
format_genecycle(ls_res_groups, w_params, added_group)
}
\arguments{
\item{ls_res_groups}{A list with results from a GeneCycle rhythmicity
analysis, split into groups.}

\item{w_params}{A data frame with results from a harmonic regression
analysis.}

\item{added_group}{If TRUE the "group" column will be removed}
}
\value{
A list of data frames containing the original and formatted results
}
\description{
Format GeneCycle Results
}
