% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_jtkcycle.R
\name{format_jtkcycle}
\alias{format_jtkcycle}
\title{Format JTK_CYCLE Results}
\usage{
format_jtkcycle(
  ls_res_groups,
  added_group,
  log_transformed = FALSE,
  log_base = 2
)
}
\arguments{
\item{ls_res_groups}{A list with results from a JTK_CYCLE rhythmicity
analysis, split into groups.}

\item{added_group}{If TRUE the "group" column will be removed}

\item{log_transformed}{Logical, if the data is log-transformed}

\item{log_base}{Logarithmic base of data. Only relevant if \code{log_transformed}
is \code{TRUE}.}
}
\value{
A list of data frames containing the original and formatted results
}
\description{
Format JTK_CYCLE Results
}
