% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_limorhyde.R
\name{format_limorhyde}
\alias{format_limorhyde}
\title{Format LimoRhyde Results}
\usage{
format_limorhyde(res_original, period)
}
\arguments{
\item{res_original}{A data frame with results from a LimoRhyde rhythmicity
analysis.}

\item{period}{Period from \code{CircadianData} object}
}
\value{
A list of data frames containing the original and formatted results
}
\description{
Format LimoRhyde Results
}
