% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_ls.R
\name{format_ls}
\alias{format_ls}
\title{Format Lomb-Scargle Results}
\usage{
format_ls(ls_res_groups, w_params, added_group, log_transformed, log_base)
}
\arguments{
\item{ls_res_groups}{A list with results from a Lomb-Scargle rhythmicity
analysis, split into groups.}

\item{w_params}{A data frame with results from a harmonic regression
analysis.}

\item{added_group}{If TRUE the "group" column will be removed}

\item{log_transformed}{Logical, if the data is log-transformed}

\item{log_base}{Logarithmic base of data. Only relevant if \code{log_transformed}
is \code{TRUE}.}
}
\value{
A list of data frames containing the original and formatted results
}
\description{
Format Lomb-Scargle Results
}
