% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_rain.R
\name{format_rain}
\alias{format_rain}
\title{Format RAIN Results}
\usage{
format_rain(ls_res_groups, added_group)
}
\arguments{
\item{ls_res_groups}{A list with results from a RAIN rhythmicity
analysis, split into groups.}

\item{added_group}{If TRUE the "group" column will be removed}
}
\value{
A list of data frames containing the original and formatted results
}
\description{
Format RAIN Results
}
