% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CircadianData.R
\name{get_wave_params}
\alias{get_wave_params}
\title{Extract Wave Parameters from a CircadianData Object}
\usage{
get_wave_params(cd)
}
\arguments{
\item{cd}{A \code{CircadianData} object.}
}
\value{
A data frame with estimated wave parameters
}
\description{
\code{get_wave_params()} retrieves the estimated wave parameters stored inside a
\code{CircadianData} object.
}
