% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalise_dataset.R
\name{normalise_dataset}
\alias{normalise_dataset}
\title{Turn RNA-seq counts to log2 CPM values}
\usage{
normalise_dataset(cd, verbose = TRUE)
}
\arguments{
\item{cd}{A \code{CircadianData} object with count data in the \code{results} slot}

\item{verbose}{Logical, whether a message should be printed}
}
\value{
The updated \code{CircadianData} object with log2 CPM values in the
\code{results} slot
}
\description{
Turn RNA-seq counts to log2 CPM values
}
