% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_circan.R
\name{run_circan}
\alias{run_circan}
\title{Run CircaN rhythmicity detection}
\usage{
run_circan(inputs, ...)
}
\arguments{
\item{inputs}{Inputs for rhythmicity detection function generated by
\code{prepare_circan()}}

\item{...}{Additional parameters passed to
\code{CircaN::circan()}}
}
\value{
A data frame
}
\description{
Run CircaN rhythmicity detection
}
