% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_limorhyde.R
\name{run_limorhyde}
\alias{run_limorhyde}
\title{Detect rhythmicity with LimoRhyde}
\usage{
run_limorhyde(cd)
}
\arguments{
\item{cd}{A \code{CircadianData} object.}
}
\value{
A data frame with the results of a LimoRhyde analysis.
}
\description{
This functions runs rhythmicity detection using LimoRhyde. It expects a
\code{CircadianDaata} object as input. The metadata slot of this object must
contain a data.frame with a column named "Time" that contains time point
information for a given sample.
}
