% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_repeatedcircadian.R
\name{run_repeatedcircadian}
\alias{run_repeatedcircadian}
\title{Run RepeatedCircadian rhythmicity detection}
\usage{
run_repeatedcircadian(inputs, ...)
}
\arguments{
\item{inputs}{Inputs for rhythmicity detection function generated by
\code{prepare_repeatedcircadian()}}

\item{...}{Additional parameters passed to
\code{RepeatedCircadian::rpt_rhythmicity()}}
}
\value{
A data frame
}
\description{
Run RepeatedCircadian rhythmicity detection
}
