% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CircadianData.R
\name{show,CircadianData-method}
\alias{show,CircadianData-method}
\title{Show Method for CircadianData}
\usage{
\S4method{show}{CircadianData}(object)
}
\arguments{
\item{object}{A \code{CircadianData} object.}
}
\description{
Show Method for CircadianData
}
